requirements.txt
psycopg2-binary
django-allauth
xhtml2pdf
python-docx